TString: Fix bug in Split()
If we previously supplied something like:

Split("TXTR", ",")

to this, it would produce a collection with the string "TXT" in it,
which is definitely not intended behavior. Instead, we can just pass the
length through to the underlying substring operation
(basic_string::substr) which treats lengths past the end as saturating.